Faulty dots data from subject 14 written to csv/h5
The timestamp is 2019_12_11_12_21
ME_Fall_2019_I.ipynb was used to:
1/ label the reproduced dots data and store it in the global dots .csv file
2/ write the labeled dots to HDF5 (with the duplicate bug). File used is
fall2019_v1.h5

All this commit was a test to see that the code worked. The actual dots
data is wrong because I didn't match the seeding in
dotsDrawableDotKinetogramDebug with dotsDrawableDotKinetogram.

So, in a nutshell, nothing from fall2019_v1.h5 nor dots_fall_2019_v1.csv
is useable for motion energy (all the dots are faulty).

New .csv and HDF5 databases to be started, from subject 14 onwards.